Fix the escaping vs. bash problem.
Apparently in Ubuntu's /bin/sh, `echo' refers to /bin/echo, which does not
support the -e flag (interprets backslash-escaped characters). In /bin/bash,
however, `echo' refers to the builtin command, which does support the -e flag.

So what it boils down to is some difference in how Mac OS X's bash and Ubuntu's
bash handle escaping, but the fact that I can't nail down exactly what that is
probably means we'll have to revisit this in the future.